# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Atlas API",
    "description": "Atlas knowledge and project-management platform REST API",
    "contact": {
      "name": "Ignacio Perez",
      "email": "ignacio@feuer.me"
    },
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "All users (active and disabled)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not a root/admin user"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pending user created with activation link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not a root/admin user"
          },
          "422": {
            "description": "Workspace not found, or role=owner"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/workspaces/{ws}/assignable-users": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Lists users who can be added to the workspace.",
        "description": "Returns the active (non-disabled) users that are NOT already members of this\nworkspace. The `users` table holds only human accounts, so api keys are\nexcluded by construction. Mirrors the `GET /v1/users` shape (`Vec<UserDto>`).",
        "operationId": "list_assignable_users",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users that can be added as members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Insufficient privileges"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/workspaces/{ws}/groups/{group_id}/members/{user_id}": {
      "delete": {
        "tags": [
          "groups"
        ],
        "operationId": "remove_group_member",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group_id",
            "in": "path",
            "description": "Group ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID to remove",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed from group"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Insufficient privileges"
          },
          "404": {
            "description": "Group or membership not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/workspaces/{ws}/members": {
      "get": {
        "tags": [
          "members"
        ],
        "operationId": "list_workspace_members",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace members and agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrincipalDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Workspace not found or principal is not a member"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "members"
        ],
        "summary": "Adds an existing user to the workspace at the requested role.",
        "description": "Mirrors the authorization discipline of `update_member_role`:\n1. `WorkspaceOwnerOrAdmin` gates structural access (resolves caller class).\n2. Role string is parsed; an unknown value → 422.\n3. Role-grant matrix: an admin may add `member` or `admin`, never `owner`;\n   only an owner (or break-glass) may add an `owner`.\n4. The target user must exist (404) and must not be disabled (422 — a\n   deactivated account cannot be assigned).\n5. The user must not already be a member → 409 Conflict.\n6. The membership insert and its audit row share one transaction.",
        "operationId": "add_member",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User added to the workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Insufficient privileges to grant the requested role"
          },
          "404": {
            "description": "Workspace or target user not found"
          },
          "409": {
            "description": "User is already a member"
          },
          "422": {
            "description": "Unknown role string, or target user is disabled"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/workspaces/{ws}/members/{user_id}": {
      "delete": {
        "tags": [
          "members"
        ],
        "summary": "Removes a member from the workspace.",
        "description": "Enforces the same ordered authorization checks as `update_member_role`, adapted for\nDELETE:\n- Step 3: admin cannot remove an owner.\n- Step 4: no one (including break-glass) can remove the last owner → 409.",
        "operationId": "remove_member",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Target member user ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Insufficient privileges"
          },
          "404": {
            "description": "Workspace or target member not found"
          },
          "409": {
            "description": "Last-owner lockout"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "members"
        ],
        "summary": "Changes a workspace member's role.",
        "description": "Enforces strictly ordered authorization checks:\n1. `WorkspaceOwnerOrAdmin` extractor gates structural access (resolves caller class).\n2. Target membership is loaded; 404 if the target is not a member.\n3. Role-permission matrix: admins cannot touch owners or promote to owner; others\n   are allowed subject to step 4.\n4. Last-owner-lockout (independent of step 3, applies to break-glass too): demoting\n   the sole remaining owner → 409. This is a data-integrity invariant, not a\n   permission — break-glass is NOT exempt.\n5. Apply `update_role`. Idempotent: same-role PATCH validates the full matrix first.\n\nSteps 3 and 4 are independent and strictly ordered. Never collapse them.",
        "operationId": "update_member_role",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "description": "Workspace slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Target member user ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON body"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Insufficient privileges"
          },
          "404": {
            "description": "Workspace or target member not found"
          },
          "409": {
            "description": "Last-owner lockout"
          },
          "422": {
            "description": "Unknown role string"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddMemberRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/workspaces/{ws}/members`.\n\nAdds an existing user to the workspace at the given role. `role` is one of\n`\"owner\"` | `\"admin\"` | `\"member\"`; granting `owner` requires an owner (or\nbreak-glass) caller.",
        "required": [
          "user_id",
          "role"
        ],
        "properties": {
          "role": {
            "type": "string"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateUserRequest": {
        "type": "object",
        "description": "Request body for `POST /v1/users`.\n\nCreates a pending account with no password. The returned `activation_link`\nmust be shared with the invitee so they can set their own password.",
        "required": [
          "username",
          "display_name",
          "workspace",
          "role"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": "string",
            "description": "Membership role: `\"admin\"` or `\"member\"` (owner is rejected with 422)."
          },
          "username": {
            "type": "string"
          },
          "workspace": {
            "type": "string",
            "description": "Workspace slug where the new user will be added."
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "description": "Public user representation (no password hash).",
        "required": [
          "id",
          "username",
          "display_name",
          "is_root",
          "is_system_admin",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "activated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "`None` means the account has not yet been activated."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "disabled_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_root": {
            "type": "boolean"
          },
          "is_system_admin": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}